Testing conversion of ImagePlus to ImgLib2
Ideally all images would use the ImgLib2 Img format.  Currently they're all stored as the old ImagePlus format.  The main problem with this is that Img requires types to be specified.  It may be easier to start by trying to convert a smaller plugin (e.g. a particular image filter plugin such as DoG).